requirements.txt
streamlit
plotly.express
requests
dotenv
os